[organization] Rename includes/ -> util/
