+++ util/bit_buffer.cpp
