+++ util/crc32.c
