+++ util/string_raw_data_encoder.c
