+++ util/string_raw_data_encoder.hpp
